Authentication & Authorization

+++ src/articles/articles.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Article, Tag } from '@prisma/client';
 
 import * as C from '../constants';
@@ -120,9 +124,17 @@ export class ArticlesService {
     return this.safeArticle(article);
   }
 
-  async update(id: string, dto: UpdateArticleDto): Promise<T.ArticleType> {
+  async update(
+    id: string,
+    dto: UpdateArticleDto,
+    user: T.TokenPayloadType,
+  ): Promise<T.ArticleType> {
     const exists = await this.prisma.article.findUnique({ where: { id } });
     if (!exists) throw new NotFoundException(C.ARTICLE_NOT_FOUND);
+
+    if (user.role === C.EDITOR && exists.authorId !== user.userId) {
+      throw new ForbiddenException(C.EDIT_EXCEPTION);
+    }
 
     const updated = await this.prisma.article.update({
       where: { id },
@@ -145,12 +157,15 @@ export class ArticlesService {
     return this.safeArticle(updated);
   }
 
-  async remove(id: string) {
+  async remove(id: string, user: T.TokenPayloadType) {
     const exists = await this.prisma.article.findUnique({ where: { id } });
     if (!exists) throw new NotFoundException(C.ARTICLE_NOT_FOUND);
 
-    await this.prisma.article.delete({ where: { id } });
+    if (user.role !== C.ADMIN && exists.authorId !== user.userId) {
+      throw new ForbiddenException(C.ARTICLE_DELETE_EXCEPTION);
+    }
 
+    await this.prisma.article.delete({ where: { id } });
     return null;
   }
 }
